requirements.txt
fastapi
uvicorn[standard]
sqlmodel
pydantic
aiofiles
python-multipart
ffmpeg-python
jinja2

